+++ Products/CMFPlone/factory.py
@@ -1,23 +1,25 @@
-from Products.CMFPlone.Portal import PloneSite
+# -*- coding: utf-8 -*-
+from plone.registry.interfaces import IRegistry
 from Products.CMFPlone.events import SiteManagerCreatedEvent
 from Products.CMFPlone.interfaces import INonInstallable
+from Products.CMFPlone.Portal import PloneSite
 from Products.GenericSetup.tool import SetupTool
-from plone.registry.interfaces import IRegistry
 from zope.component import queryUtility
 from zope.event import notify
-from zope.interface import implements
+from zope.interface import implementer
 from zope.site.hooks import setSite
 
 _TOOL_ID = 'portal_setup'
 _DEFAULT_PROFILE = 'Products.CMFPlone:plone'
+_TYPES_PROFILE = 'plone.app.contenttypes:default'
 _CONTENT_PROFILE = 'plone.app.contenttypes:plone-content'
 
 # A little hint for PloneTestCase
 _IMREALLYPLONE5 = True
 
 
+@implementer(INonInstallable)
 class HiddenProfiles(object):
-    implements(INonInstallable)
 
     def getNonInstallableProfiles(self):
         return [_DEFAULT_PROFILE,
@@ -72,11 +74,20 @@ def zmi_constructor(context):
     return request.response.redirect(url + '/@@plone-addsite?site_id=Plone')
 
 
-def addPloneSite(context, site_id, title='Plone site', description='',
-                 profile_id=_DEFAULT_PROFILE,
-                 content_profile_id=_CONTENT_PROFILE, snapshot=False,
-                 extension_ids=(), setup_content=True,
-                 default_language='en', portal_timezone='UTC'):
+def addPloneSite(
+    context,
+    site_id,
+    title='Plone site',
+    description='',
+    profile_id=_DEFAULT_PROFILE,
+    types_profile_id=_TYPES_PROFILE,
+    content_profile_id=_CONTENT_PROFILE,
+    snapshot=False,
+    extension_ids=(),
+    setup_content=True,
+    default_language='en',
+    portal_timezone='UTC'
+):
     """Add a PloneSite to the context."""
     context._setObject(site_id, PloneSite(site_id))
     site = context._getOb(site_id)
@@ -97,23 +108,28 @@ def addPloneSite(context, site_id, title='Plone site', description='',
     reg['plone.default_language'] = default_language
     reg['plone.available_languages'] = [default_language]
 
-    if setup_content:
+    # setup content types
+    if not setup_content and types_profile_id is not None:
         setup_tool.runAllImportStepsFromProfile(
-            'profile-%s' % content_profile_id)
+            'profile-{0:s}'.format(types_profile_id)
+        )
+    # setup initial (example) content (with content types)
+    if setup_content and content_profile_id is not None:
+        setup_tool.runAllImportStepsFromProfile(
+            'profile-{0:s}'.format(content_profile_id)
+        )
 
-    props = dict(
-        title=title,
-        description=description,
-    )
-    # Do this before applying extension profiles, so the settings from a
-    # properties.xml file are applied and not overwritten by this
-    site.manage_changeProperties(**props)
+    # Set portals title and description before applying extension profiles,
+    # so the settings from a properties.xml file are applied and not
+    # overwritten by this
+    site.manage_changeProperties(title=title, description=description)
 
+    # apply extension profiles
     for extension_id in extension_ids:
         setup_tool.runAllImportStepsFromProfile(
             'profile-%s' % extension_id)
 
-    if snapshot is True:
+    if snapshot:
         setup_tool.createSnapshot('initial_configuration')
 
     return site